refactor: Move MapStyleConfigurator component to its own file
Extracts LayerRow, MapStyleConfiguratorDialog, and MapStyleConfiguratorOpComponent
from op-components.tsx into map-style-configurator-op.tsx. Exports the shared
node helpers (useLocked, useNodeDimmed, useFieldVisibility, NodeHeader, OutputHandle)
needed by the new file.
